+++ website/src/examples/general_form.rs
@@ -2,7 +2,9 @@ use leptos::ev::SubmitEvent;
 use leptos::html::{Input, Select};
 use leptos::*;
 use leptos_bulma::elements::BButton;
-use leptos_bulma::form::{BFile, BPasswordField, BSelectField, BTextField, BTextareaField};
+use leptos_bulma::form::{
+    BCheckboxField, BFile, BPasswordField, BSelectField, BTextField, BTextareaField,
+};
 
 fn validate_required(value: Option<String>, error: RwSignal<Option<String>>) -> bool {
     let value_is_empty = value.unwrap_or_default().trim().is_empty();
@@ -13,7 +15,20 @@ fn validate_required(value: Option<String>, error: RwSignal<Option<String>>) -> 
         None
     });
 
-    value_is_empty
+    !value_is_empty
+}
+
+fn validate_required_checkbox(node_ref: NodeRef<Input>, error: RwSignal<Option<String>>) -> bool {
+    validate_required(
+        node_ref.get().and_then(|el| {
+            if el.checked() {
+                Some("checked".to_owned())
+            } else {
+                None
+            }
+        }),
+        error,
+    )
 }
 
 fn validate_required_input(node_ref: NodeRef<Input>, error: RwSignal<Option<String>>) -> bool {
@@ -27,21 +42,37 @@ fn validate_required_select(node_ref: NodeRef<Select>, error: RwSignal<Option<St
 #[component]
 pub fn GeneralForm() -> impl IntoView {
     let node_ref_first_name = create_node_ref::<Input>();
+    let node_ref_slug = create_node_ref::<Input>();
     let node_ref_email = create_node_ref::<Input>();
     let node_ref_password = create_node_ref::<Input>();
     let node_ref_country = create_node_ref::<Select>();
+    let node_ref_accept_terms = create_node_ref::<Input>();
     let error_first_name = create_rw_signal(None);
+    let error_slug = create_rw_signal(None);
     let error_email = create_rw_signal(None);
     let error_password = create_rw_signal(None);
     let error_country = create_rw_signal(None);
+    let error_accept_terms = create_rw_signal(None);
+    let value_slug = create_rw_signal("".to_owned());
+
+    let first_name_on_input = move |event| {
+        value_slug.set(slug::slugify(event_target_value(&event)));
+        validate_required_input(node_ref_first_name, error_first_name);
+    };
 
     let form_on_submit = move |event: SubmitEvent| {
         event.prevent_default();
 
-        validate_required_input(node_ref_first_name, error_first_name);
-        validate_required_input(node_ref_email, error_email);
-        validate_required_input(node_ref_password, error_password);
-        validate_required_select(node_ref_country, error_country);
+        let is_valid = validate_required_input(node_ref_first_name, error_first_name)
+            & validate_required_input(node_ref_slug, error_slug)
+            & validate_required_input(node_ref_email, error_email)
+            & validate_required_input(node_ref_password, error_password)
+            & validate_required_select(node_ref_country, error_country)
+            & validate_required_checkbox(node_ref_accept_terms, error_accept_terms);
+
+        if is_valid {
+            let _ = window().alert_with_message("Just an example, nothing is sent.");
+        }
     };
 
     view! {
@@ -50,8 +81,16 @@ pub fn GeneralForm() -> impl IntoView {
                 node_ref=node_ref_first_name
                 label="* First name"
                 error=error_first_name
+                on_input=first_name_on_input
+            />
+
+            <BTextField
+                node_ref=node_ref_slug
+                label="* Slug"
+                value=value_slug
+                error=error_slug
                 on_input=move |_| {
-                    validate_required_input(node_ref_first_name, error_first_name);
+                    validate_required_input(node_ref_slug, error_slug);
                 }
             />
 
@@ -94,6 +133,15 @@ pub fn GeneralForm() -> impl IntoView {
             />
 
             <BFile label="Avatar image" accept="image/gif,image/jpeg,image/png"/>
+            <BCheckboxField
+                node_ref=node_ref_accept_terms
+                label="I accept the terms and conditions"
+                error=error_accept_terms
+                on_change=move |_| {
+                    validate_required_checkbox(node_ref_accept_terms, error_accept_terms);
+                }
+            />
+
             <BButton class="is-primary">"Submit"</BButton>
         </form>
     }