+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Checkbox Leptos component.
+
 ## [0.0.0-alpha.9] - 2024-04-10
 
 # Added